% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{agreement}
\alias{agreement}
\title{Model agreement}
\usage{
agreement(array3d, threshold)
}
\arguments{
\item{array3d}{3d array in which first dimension is model member, usually C4R$Data. This can be obtained as a result of
bindGrid(data, dimension = "member"). Temporal dimension needs to be removed through another function}

\item{threshold}{numeric (0.5-1). Percentage of model agreement required}
}
\value{
array with 0 (no model agreement) or 1 (model agreement), based on threshold
}
\description{
function to find models members sign agreement.
}
