% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{array.lme}
\alias{array.lme}
\title{Apply mixed effect model to multidimensional array}
\usage{
array.lme(cl4, slope)
}
\arguments{
\item{cl4}{list used in climate4R with slot Data containing more than one member}

\item{slope}{logical.}
}
\description{
fit a mixed effect model (random slope) to c4R object with multiple member in the first dimension
}
