% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavaR.R
\docType{package}
\name{CAVAanalytics}
\alias{CAVAanalytics}
\title{CAVAanalytics Package}
\description{
Contains functions to load and visualize climate models
}
\author{
Riccardo Soldan \email{riccardosoldan@hotmail.it}
}
