% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{common_dates}
\alias{common_dates}
\title{Date selection}
\usage{
common_dates(data)
}
\arguments{
\item{data}{list containing C4R objects, which are the outputs of the loadGridata function}
}
\description{
automatically select common dates among C4R objects
}
