% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{CONSTANTS}
\alias{CONSTANTS}
\title{Constants and configurations}
\format{
An object of class \code{list} of length 2.
}
\usage{
CONSTANTS
}
\description{
Constants and configurations
}
\keyword{internal}
