% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_vector_to_month_initials}
\alias{convert_vector_to_month_initials}
\title{Convert month to initials}
\usage{
convert_vector_to_month_initials(month_vector)
}
\arguments{
\item{month_vector}{numeric vector, 1 to 12.}
}
\description{
Convert numeric vector of months into month initials
}
