% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ens_trends}
\alias{ens_trends}
\title{Apply multivariate linear regression to a multimember grid}
\usage{
ens_trends(c4R)
}
\value{
array, without spatial averages, or dataframe, if spatial averages are performed
}
\description{
This function can be used after performing annual aggregation and with a multigrid object. it applies multivariate linear regression per pixel if
spatial averages are not performed or for spatially aggregated data
}
