% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raster.R
\name{extract_raster}
\alias{extract_raster}
\title{Write rasters}
\usage{
extract_raster(step2, filename, ensemble)
}
\arguments{
\item{step2}{output of one of CAVAanalytics functions, such as projections.}

\item{filename}{charachter, specifying the file name (needs to end in .tif)}

\item{ensemble}{logical. Whether to save the raster for the ensemble or for individual models}

\item{stat}{charachter. One of mean or sd. Only valid when ensemble is TRUE}
}
\description{
Automatically write rasters from CAVAanalytics step two
}
