% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{IPCC_palette}
\alias{IPCC_palette}
\title{IPCC color palette}
\usage{
IPCC_palette(type, divergent)
}
\arguments{
\item{type}{character, one of tmp or pr.}

\item{divergent}{logical. If TRUE, divergent palette are used. Useful in combination with legend.range to assign central colors in the palette to zero values}
}
\description{
automatically use the suggested IPCC colors for precipitation and temperature.
}
