% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_and_trends.R
\name{load_data_and_trends}
\alias{load_data_and_trends}
\title{Load data and apply trends in spatial chunks}
\usage{
load_data_and_trends(
  variable,
  years.hist = NULL,
  years.proj,
  path.to.data,
  path.to.obs = NULL,
  xlim,
  ylim,
  aggr.m = "none",
  domain = NULL,
  chunk.size,
  overlap = 0.5,
  season,
  lowert = NULL,
  uppert = NULL,
  consecutive = F,
  scaling.type = "additive",
  duration = "max",
  bias.correction = F,
  intraannual_var = F,
  n.sessions = 1
)
}
\arguments{
\item{variable}{character, indicating variable name}

\item{years.hist}{Numerical range, years to select for historical simulations and observations}

\item{years.proj}{Numerical range, years to select for projections}

\item{path.to.data}{path to the directory containing the RCP/SSPs folders and historical simulations (optional). For example,
home/user/data/. data would contain subfolders with the climate/impact models. Historical simulations have to be contained in a folder called historical. If path.to.data is set as CORDEX-CORE, CORDEX-CORE simulations from RCM RegCM4 will be loaded}

\item{path.to.obs}{Default to NULL, if not, indicate the absolute path to the directory containing a reanalysis dataset, for example ERA5. To automatically load W5E5. specify W5E5}

\item{xlim}{numeric of length = 2, with minimum and maximum longitude coordinates, in decimal degrees, of the bounding box selected.}

\item{ylim}{same as xlim, but for the selection of the latitudinal range.}

\item{aggr.m}{character, monthly aggregation. One of none, mean or sum}

\item{domain}{specify the CORDEX-CORE domain (e.g AFR-22, EAS-22). Used with path.to.data = CORDEX-CORE. Default is NULL}

\item{chunk.size}{numeric, indicating the number of chunks. The smaller the better when working with limited RAM}

\item{overlap}{numeric, amount of overlap needed to create the composite. Default 0.5}

\item{season}{numeric, seasons to select. For example, 1:12}

\item{lowert}{numeric of length=1, lower threshold}

\item{uppert}{numeric of length=1, upper threshold}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{character, either "max" or "total"}

\item{bias.correction}{logical}

\item{intraannual_var}{logical, whether linear regression is applied to annual variability, measured as standard deviation}

\item{n.sessions}{numeric, number of sessions to use, default is one. Parallelisation can be useful when multiple scenarios are used (RCPS, SSPs). However, note that parallelising will increase RAM usage}
}
\value{
list with merged raster stacks. To explore the output run attributes(output)
}
\description{
Automatically load and process climate model projections in a memory efficient way. Useful for analysing large areas. Arguments historical and country are not available for this function.
}
