% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_hub.R
\name{load_data_hub}
\alias{load_data_hub}
\title{Models upload (only works from the University of Cantabria Jupyter HUB environment)}
\usage{
load_data_hub(
  path.to.data,
  country,
  variable,
  xlim = NULL,
  ylim = NULL,
  years.hist = NULL,
  years.proj,
  path.to.obs = NULL,
  buffer = 0,
  domain = NULL,
  aggr.m = "none",
  n.sessions = 6,
  years.obs = NULL
)
}
\arguments{
\item{path.to.data}{character, indicating the database of interest (e.g CORDEX-CORE).}

\item{country}{character, in English, indicating the country of interest or an object of class sf. To select a bounding box,
set country to NULL and define arguments xlim and ylim}

\item{variable}{character indicating the variable name}

\item{xlim}{numeric of length = 2, with minimum and maximum longitude coordinates, in decimal degrees, of the bounding box of interest}

\item{ylim}{same as xlim, but for the selection of the latitudinal range}

\item{years.hist}{numeric, specify year range for the historical experiment}

\item{years.proj}{numeric, specify year range for projections}

\item{path.to.obs}{character, default to NULL. To automatically load W5E5 or ERA5, specify W5E5 or ERA5.}

\item{buffer}{numeric, default is zero.}

\item{domain}{charachter, specify the CORDEX-CORE domain (e.g AFR-22, EAS-22).}

\item{aggr.m}{character, monthly aggregation. One of none, mean or sum}

\item{n.sessions}{numeric, number of sessions to use in parallel processing. Default to 6. Increasing the number of sessions will not necessarily results in better performances. Leave as default unless necessary}

\item{years.obs}{NULL or numeric, specify year range for observation. Specifying years.obs will overwrite years.hist for observations}
}
\value{
list of length 2. List[[1]] contains a tibble with list columns and List[[2]] the bbox
}
\description{
Automatically upload databases available at UC servers
}
