% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_hub.R
\name{load_model_data.hub}
\alias{load_model_data.hub}
\title{Load model data from specified source}
\usage{
load_model_data.hub(
  x,
  variable,
  years,
  xlim,
  ylim,
  aggr.m,
  database,
  temporal_chunking
)
}
\description{
Load model data from specified source
}
\keyword{internal}
