% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_model_data}
\alias{load_model_data}
\title{Load model data from specified source}
\usage{
load_model_data(
  x,
  variable,
  years,
  xlim,
  ylim,
  aggr.m,
  path.to.data,
  temporal_chunking
)
}
\description{
Load model data from specified source
}
\keyword{internal}
