% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_hub.R
\name{load_observation_data.hub}
\alias{load_observation_data.hub}
\title{Load observational data}
\usage{
load_observation_data.hub(
  obs.file,
  variable,
  years,
  xlim,
  ylim,
  aggr.m,
  path.to.obs
)
}
\description{
Load observational data
}
\keyword{internal}
