% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_observation_data}
\alias{load_observation_data}
\title{Load observational data}
\usage{
load_observation_data(
  obs.file,
  variable,
  years,
  xlim,
  ylim,
  aggr.m,
  path.to.obs,
  temporal_chunking,
  temporal_chunk_size
)
}
\description{
Load observational data
}
\keyword{internal}
