% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadGridData_xarray}
\alias{loadGridData_xarray}
\title{Load data using xarray}
\usage{
loadGridData_xarray(dataset, var, lonLim, latLim, years, aggr.m = "none")
}
\description{
automatically using xarray and converting into C4R list for remote data access. Only works for data stored at UC server
}
