% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_raster}
\alias{make_raster}
\title{make a raster}
\usage{
make_raster(cl4.object)
}
\arguments{
\item{cl4.object}{A C4R list with the Data slot in two dimension}
}
\value{
spatRaster
}
\description{
Make a raster from a C4R list with dim(Data)=2
}
