% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_biases.R
\name{model_biases}
\alias{model_biases}
\title{Visualization  of model biases}
\usage{
model_biases(
  data,
  bias.correction = F,
  uppert = NULL,
  lowert = NULL,
  season,
  consecutive = F,
  n.sessions = 1,
  duration = "max"
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{numeric, seasons to select. For example, 1:12}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{n.sessions}{numeric, number of sessions to use, default is one. Parallelization can be useful when multiple scenarios are used (RCPS, SSPs). However, note that parallelizing will increase RAM usage}

\item{duration}{character, either "max" or "total". Used only when consecutive is TRUE}
}
\value{
list with SpatRaster. To explore the output run attributes(output)
}
\description{
Automatically compute the difference between observations and the historical experiment
}
