% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{models_trends}
\alias{models_trends}
\title{Apply linear regression to each member of multimember grid}
\usage{
models_trends(c4R, historical = F)
}
\value{
array, without spatial averages, or dataframe, if spatial averages are performed
}
\description{
This function can be used after performing annual aggregation and with a multigrid object. it applies linear regression per pixel if
spatial averages are not performed or for spatially aggregated data.
}
