% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\alias{plotting.CAVAanalytics_projections}
\alias{plotting.CAVAanalytics_ccs}
\alias{plotting.CAVAanalytics_observations}
\alias{plotting.CAVAanalytics_model_biases}
\title{Plot CAVAanalytics Results}
\usage{
plotting(rst, palette, ...)

\method{plotting}{CAVAanalytics_projections}(
  rst,
  palette = NULL,
  legend_range = NULL,
  plot_titles = "default",
  ensemble = TRUE,
  bins = FALSE,
  intervals = NULL,
  alpha = DEFAULT_ALPHA,
  stat = "mean",
  spatiotemporal = FALSE,
  temporal = FALSE,
  lwd = DEFAULT_LINE_WIDTH,
  n.groups = 3
)

\method{plotting}{CAVAanalytics_ccs}(
  rst,
  palette = NULL,
  legend_range = NULL,
  plot_titles = "default",
  ensemble = TRUE,
  bins = FALSE,
  intervals = NULL,
  alpha = DEFAULT_ALPHA,
  stat = "mean",
  spatiotemporal = FALSE,
  temporal = FALSE,
  lwd = DEFAULT_LINE_WIDTH
)

\method{plotting}{CAVAanalytics_observations}(
  rst,
  palette = NULL,
  legend_range = NULL,
  plot_titles = "default",
  ensemble = FALSE,
  bins = FALSE,
  intervals = NULL,
  alpha = DEFAULT_ALPHA,
  spatiotemporal = FALSE,
  temporal = FALSE,
  lwd = DEFAULT_LINE_WIDTH,
  n.groups = 3
)

\method{plotting}{CAVAanalytics_model_biases}(
  rst,
  palette = NULL,
  legend_range = NULL,
  plot_titles = "default",
  ensemble = TRUE,
  bins = FALSE,
  intervals = NULL,
  alpha = DEFAULT_ALPHA,
  temporal = FALSE,
  spatiotemporal = FALSE,
  lwd = DEFAULT_LINE_WIDTH
)
}
\arguments{
\item{rst}{Output from CAVAanalytics functions}

\item{palette}{Color palette to use. Default is NULL}

\item{...}{Additional arguments passed to specific methods}

\item{legend_range}{Fix legend limits. Default is NULL}

\item{plot_titles}{Title of the plot legend. Default is "default"}

\item{ensemble}{Whether to visualize ensemble mean or individual models. Default is TRUE}

\item{bins}{Whether to visualize colors as gradient or bins. Default is FALSE}

\item{intervals}{Control number of bins when bins=TRUE. Default is NULL}

\item{alpha}{Transparency of colors. Default is NA}

\item{stat}{Statistic to compute ("mean" or "sd"). Default is "mean"}

\item{spatiotemporal}{Whether to show frequencies without aggregation. Default is FALSE}

\item{temporal}{Whether to show temporal aggregation. Default is FALSE}

\item{lwd}{Width of country boundaries. Default is 0.1}

\item{n.groups}{Number of groups for spatiotemporal plots. Default is 3}
}
\value{
A ggplot2 object
}
\description{
Generic plotting function for CAVAanalytics outputs
}
\details{
Plot CAVAanalytics Results
}
