% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{remove_facets}
\alias{remove_facets}
\title{Clean text from plots}
\usage{
remove_facets(position = "both")
}
\arguments{
\item{position}{character indicating which facets to remove (both, x or y).}
}
\description{
Remove facets labels from ggplot object
}
