% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rename_facets}
\alias{rename_facets}
\title{Customise facet text}
\usage{
rename_facets(p, current_label, new_label, position = c("x", "y"))
}
\arguments{
\item{p}{Output of plotting function}

\item{current_label}{character indicating the current facet label.}

\item{new_label}{character indicating which new label to use.}

\item{position}{character indicating which facets to change (x or y).}
}
\description{
Customise facets labels in ggplot object. It only works when ensemble is TRUE
}
