% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thrs}
\alias{thrs}
\title{Calculation of thresholds}
\usage{
thrs(col, lowert, uppert, ...)
}
\arguments{
\item{col}{numeric vector}

\item{lowert}{numeric. lower threshold}

\item{uppert}{numeric. upper threshold}
}
\value{
numeric of length 1
}
\description{
Calculation of number of days with certain condition. It can be used with aggregateGrid.
}
