% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thrs_consec}
\alias{thrs_consec}
\title{Consecutive days}
\usage{
thrs_consec(col, duration, lowert, uppert, frequency)
}
\arguments{
\item{col}{numeric vector}

\item{duration}{either "max" or "total".}

\item{lowert}{numeric. Lower threshold}

\item{uppert}{numeric. Upper threshold}

\item{frequency}{logical. Whether frequency or abosulte numbers should be returned. Only works with duration != max}
}
\value{
numeric of length 1
}
\description{
Calculation of consecutive days. It can be used with aggregateGrid.
}
