% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends}
\alias{trends}
\title{Analysis of long-term trends}
\usage{
trends(
  data,
  bias.correction = FALSE,
  uppert = NULL,
  lowert = NULL,
  season,
  consecutive = FALSE,
  duration = "max",
  intraannual_var = FALSE,
  observation = FALSE,
  n.sessions = 1
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical, whether to perform bias.correction or not}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{list, containing seasons to select. For example, list(1:6, 7:12)}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{either "max" or specify a number. Used only when consecutive is TRUE. For example, to know the number of consecutive days with tmax above 35, lasting more than 3 days, specify uppert=35, consecutive =T and duration=3}

\item{intraannual_var}{logical, whether linear regression is applied to annual variability, measured as standard deviation}

\item{observation}{logical, whether to visualize trends for the observational dataset or projections}

\item{n.sessions}{numeric, number of sessions to use, default is one. Parallelisation can be useful when multiple scenarios are used (RCPS, SSPs). However, note that parallelising will increase RAM usage}
}
\value{
list with SpatRaster. To explore the output run attributes(output)
}
\description{
Compute multivariate linear regression and linear regression through design-based inference
}
