% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{years_selection}
\alias{years_selection}
\title{Subset years in CAVAlist}
\usage{
years_selection(CAVAlist, years, projections = T)
}
\arguments{
\item{CAVAlist}{output of the load_data or load_data_hub function}

\item{years}{numeric, years to select}

\item{projections}{logical, whether years selection should be applied to projections or observations}
}
\description{
Select specific years for projections or observations in a CAVAlist
}
