% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_change_signal.R
\name{climate_change_signal}
\alias{climate_change_signal}
\title{Calculation of climate change signal}
\usage{
climate_change_signal(
  data,
  uppert = NULL,
  lowert = NULL,
  season,
  consecutive = F,
  duration = "max",
  bias.correction = F,
  n.sessions = 1
)
}
\arguments{
\item{data}{output of load_data}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{list, containing seasons to select. For example, list(1:6, 7:12)}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{either "max" or specify a number. Used only when consecutive is TRUE. For example, to know the number of consecutive days with tmax above 35, lasting more than 3 days, specify uppert=35, consecutive =T and duration=3}

\item{bias.correction}{logical}

\item{n.sessions}{numeric, number of sessions to use, default is one. Parallelisation can be useful when multiple scenarios are used (RCPS, SSPs). However, note that parallelising will increase RAM usage}
}
\value{
list with SpatRaster. To explore the output run attributes(output)
}
\description{
Automatically computes climate change signal
}
