% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_change_signal.R
\name{climate_change_signal}
\alias{climate_change_signal}
\title{climate change signal analysis}
\usage{
climate_change_signal(
  data,
  uppert = NULL,
  lowert = NULL,
  season,
  consecutive = F,
  duration = "max",
  bias.correction = F,
  scaling.type = "additive"
)
}
\arguments{
\item{data}{output of load_data}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{Numerical, seasons to select. For example, 1:12}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{character, either "max" or "total"}

\item{bias.correction}{logical}

\item{scaling.type}{character, default to "additive". Indicates whether to use multiplicative or additive approach for bias correction}
}
\value{
list with raster stacks
}
\description{
Automatically computes climate change signal
}
\examples{
load_data(country = "Somalia", variable="tas", years.hist=2000, years.proj=2010,
              path.to.data = "CORDEX-CORE", domain="AFR-22") \%>\%
climate_change_signal(., season = 1:12)
}
