% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_climate_change_signal.R
\name{load_data_and_climate_change_signal}
\alias{load_data_and_climate_change_signal}
\title{Load data and apply climate change signal in spatial chunks}
\usage{
load_data_and_climate_change_signal(
  variable,
  years.hist = NULL,
  years.proj,
  path.to.data,
  path.to.obs = NULL,
  xlim,
  ylim,
  aggr.m = "none",
  domain,
  chunk.size,
  overlap = 0.5,
  season,
  lowert = NULL,
  uppert = NULL,
  consecutive = F,
  scaling.type = "additive",
  duration = "max",
  bias.correction = F,
  n.sessions = 1
)
}
\arguments{
\item{variable}{character indicating the variable name}

\item{years.hist}{numeric, specify year range for the historical experiment}

\item{years.proj}{numeric, specify year range for projections}

\item{path.to.data}{path to the directory containing the RCP/SSPs folders and historical simulations (optional). For example,
home/user/data/. data would contain subfolders with the climate/impact models. Historical simulations have to be contained in a folder called historical. If path.to.data is set as CORDEX-CORE, CORDEX-CORE simulations will be loaded}

\item{path.to.obs}{character, default to NULL. To automatically load W5E5 or ERA5, specify W5E5 or ERA5. Otherwise, indicate the absolute path to the directory containing another observational dataset}

\item{xlim}{numeric of length = 2, with minimum and maximum longitude coordinates, in decimal degrees, of the bounding box of interest}

\item{ylim}{same as xlim, but for the selection of the latitudinal range}

\item{aggr.m}{character, monthly aggregation. One of none, mean or sum}

\item{domain}{charachter, specify the CORDEX-CORE domain (e.g AFR-22, EAS-22). Used with path.to.data = CORDEX-CORE. Default is NULL}

\item{chunk.size}{numeric, indicating the number of chunks to. The smaller the better when working with limited RAM.}

\item{overlap}{numeric, amount of overlap needed to create the composite. Default 0.5}

\item{season}{numeric, seasons to select. For example, 1:12}

\item{lowert}{numeric of length=1, lower threshold}

\item{uppert}{numeric of length=1, upper threshold}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{character, either "max" or "total"}

\item{bias.correction}{logical}

\item{n.sessions}{numeric, number of sessions to use in parallel processing. Default to 6. Increasing the number of sessions will not necessarily results in better performances. Leave as default unless necessary}
}
\value{
list with SpatRasters. .[[1]] contains SpatRaster for the ensemble mean. .[[2]] contains SpatRaster for the ensemble sd and .[[3]] contains SpatRaster for individual models
}
\description{
Automatically load and process climate models in a memory efficient way. Useful for analysing large areas
}
\examples{
out <- load_data_and_climate_change_signal(variable="tas", years.hist=2000:2005, years.proj=2010:2015,
     path.to.data = "CORDEX-CORE", season=1:12, domain="AFR-22", xlim=c(0, 10), ylim=c(5, 10), chunk.size=2)
}
