% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_hub.R
\name{load_data_hub}
\alias{load_data_hub}
\title{Models upload (only works from the University of Cantabria Jupyter HUB environment)}
\usage{
load_data_hub(
  database = "CORDEX-CORE",
  country,
  variable,
  xlim = NULL,
  ylim = NULL,
  years.hist = NULL,
  years.proj,
  path.to.obs = NULL,
  buffer = 0,
  domain = NULL,
  aggr.m = "none",
  n.sessions = 6,
  years.obs = NULL,
  res_folder = "interp025",
  temporal_chunking = FALSE
)
}
\arguments{
\item{database}{character, indicating the database of interest (default to CORDEX-CORE). Can be CORDEX-CORE or CORDEX-CORE-BC.}

\item{country}{character, in English, indicating the country of interest or an object of class sf.}

\item{variable}{character indicating the variable name}

\item{xlim}{numeric of length = 2, with minimum and maximum longitude coordinates}

\item{ylim}{same as xlim, but for the selection of the latitudinal range}

\item{years.hist}{numeric, specify year range for the historical experiment}

\item{years.proj}{numeric, specify year range for projections}

\item{path.to.obs}{character, default to NULL. To automatically load W5E5 or ERA5}

\item{buffer}{numeric, default is zero}

\item{domain}{character, specify the CORDEX-CORE or CORDEX-CORE-BC domain}

\item{aggr.m}{character, monthly aggregation. One of none, mean or sum}

\item{n.sessions}{numeric, number of sessions for parallel processing}

\item{years.obs}{NULL or numeric, specify year range for observation}

\item{res_folder}{character, specify the resolution of the CORDEX data. Default to "interp025". Meaningful only when working with CORDEX-CORE. In the future this option will be removed}

\item{temporal_chunking}{logical, default to FALSE. If TRUE, loads data in 10-year temporal chunks}
}
\value{
list of length 2. List[[1]] contains a tibble with list columns and List[[2]] the bbox
}
\description{
Automatically upload databases available at UC servers
}
