% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_raster}
\alias{make_raster}
\title{make a raster}
\usage{
make_raster(cl4.object, dimensions, shape.file, stat = "mean")
}
\arguments{
\item{cl4.object}{A C4R list}

\item{dimensions}{vector specifying which dimensions corresponds to lat and lon. The rest will be averaged}

\item{shape.file}{sf object for which to crop and mask the spatRaster}

\item{stat}{statistic to apply. Default is mean}
}
\value{
spatRaster
}
\description{
Make a spatRaster from a C4R list
}
