% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_biases.R
\name{model_biases}
\alias{model_biases}
\title{Visualization  of model biases}
\usage{
model_biases(
  data,
  bias.correction = F,
  uppert = NULL,
  lowert = NULL,
  season,
  consecutive = F,
  frequency = F,
  n.sessions = 1,
  duration = "max"
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{list, containing seasons to select. For example, list(1:6, 7:12)}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{frequency}{logical. Used only when consecutive is TRUE and duration is not "max". For example, to know the number of heatwaves defined as the number of days with Tmax higher than 35 for at least 3 consecutive days, specify uppert=35, consecutive =T and duration=3, frequency=T}

\item{n.sessions}{numeric, number of sessions to use, default is one. Parallelization can be useful when multiple scenarios are used (RCPS, SSPs). However, note that parallelizing will increase RAM usage}

\item{duration}{either "max" or specify a number. Used only when consecutive is TRUE. For example, to know the number of consecutive days with tmax above 35, lasting more than 3 days, specify uppert=35, consecutive =T and duration=3}
}
\value{
list with SpatRaster. To explore the output run attributes(output)
}
\description{
Automatically compute the difference between observations and the historical experiment
}
