% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observations}
\alias{observations}
\title{Analysis of observations-historical period}
\usage{
observations(
  data,
  uppert = NULL,
  lowert = NULL,
  season,
  consecutive = F,
  duration = "max"
)
}
\arguments{
\item{data}{output of load_data}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{numeric, seasons to select. For example, 1:12}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{character, either "max" or "total". Used only when consecutive is TRUE}
}
\value{
list with SpatRaster. .[[1]]
}
\description{
Automatically process the observational period and compute user-defined indicators
}
