% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\title{Visualize results}
\usage{
plotting(
  rst,
  palette,
  legend_range,
  plot_titles,
  ensemble,
  bins,
  n.bins,
  alpha,
  ...
)
}
\arguments{
\item{rst}{output of one of CAVAanalytics functions, such as projections. rst is of class SpatRaster}

\item{palette}{charachter. Color Palette}

\item{legend_range}{numeric. Fix legend limits}

\item{plot_titles}{character. Title of the plot legend}

\item{ensemble}{logical. Whether to visualize the ensemble mean or each individual model}

\item{bins}{logical. Whether to visualize colors as a gradient or in bins}

\item{n.bins}{numeric. Controlling the number of bins when bins equal TRUE}

\item{alpha}{numeric. Transparency of colors}
}
\value{
ggplot object
}
\description{
Automatically plot results from CAVAanalytics main functions
}
\examples{
load_data(country = "Somalia", variable="tas", years.hist=2000, years.proj=2010,
              path.to.data = "CORDEX-CORE", domain="AFR-22") \%>\%
projections(., season = 1:12) \%>\%
plotting(plot_titles="hurs", ensemble=T)
}
