% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_three.R
\name{plotting}
\alias{plotting}
\title{Step three: Visualization}
\usage{
plotting(rst, palette = NULL, legend.range = NULL, plot_titles)
}
\arguments{
\item{rst}{output of one of cavaR functions, such as projections. rst is of class RasterStack}

\item{palette}{charachter. Color Palette}

\item{legend.range}{numeric. Fix legend limits}

\item{plot_titles}{character. Title of teh plot legend}
}
\value{
ggplot object
}
\description{
Automatically plot results from cavaR step two
}
\examples{
fpath <- system.file("extdata/", package="chatR")
exmp <- load_data(country = "Moldova", variable="hurs", n.cores=6,
              path.to.rcps = fpath) \%>\%
projections(., season = 1:12) \%>\%
plotting(plot_titles="hurs")


}
