% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\title{Visualize results}
\usage{
plotting(
  rst,
  palette,
  legend_range,
  plot_titles,
  ensemble,
  bins,
  intervals,
  alpha,
  spatiotemporal,
  temporal,
  ...
)
}
\arguments{
\item{rst}{output of one of CAVAanalytics functions, such as projections. rst is of class SpatRaster}

\item{palette}{charachter. Color Palette}

\item{legend_range}{numeric. Fix legend limits}

\item{plot_titles}{character. Title of the plot legend}

\item{ensemble}{logical. Whether to visualize the ensemble mean or each individual model}

\item{bins}{logical. Whether to visualize colors as a gradient or in bins}

\item{intervals}{numeric vector. Controlling the number of bins when bins equal TRUE}

\item{alpha}{numeric. Transparency of colors}

\item{spatiotemporal}{logical. Whether computed yearly data should be visualized without spatial and temporal aggregation. Basically, frequencies are visualized}

\item{temporal}{logical. Whether computed yearly data should be visualized temporally after spatial aggregation (median of all pixels)}
}
\value{
ggplot object
}
\description{
Automatically plot results from CAVAanalytics main functions
}
