% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_three.R
\name{plotting}
\alias{plotting}
\title{Step three: Visualization}
\usage{
plotting(rst, palette, legend.range, plot_titles, ensemble)
}
\arguments{
\item{rst}{output of one of cavaR functions, such as projections. rst is of class RasterStack}

\item{palette}{charachter. Color Palette}

\item{legend.range}{numeric. Fix legend limits}

\item{plot_titles}{character. Title of the plot legend}

\item{ensemble}{logical. Whether to visualize the ensemble mean or each individual model}
}
\value{
ggplot object
}
\description{
Automatically plot results from cavaR step two
}
\examples{
load_data(country = "Somalia", variable="tas", years.hist=2000, years.proj=2010,
              path.to.data = "CORDEX-CORE", domain="AFR-22") \%>\%
projections(., season = 1:12) \%>\%
plotting(plot_titles="hurs", ensemble=T)
}
