% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{projections}
\alias{projections}
\title{Climate projections visualization}
\usage{
projections(
  data,
  bias.correction = F,
  uppert = NULL,
  lowert = NULL,
  season,
  scaling.type = "additive",
  consecutive = FALSE,
  duration = "max"
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{Numerical, seasons to select. For example, 1:12}

\item{scaling.type}{character, default to "additive". Indicates whether to use multiplicative or additive approach for bias correction}

\item{consecutive}{logical, to use in conjuunction with lowert or uppert}

\item{duration}{character, either "max" or "total".}
}
\value{
raster
}
\description{
Automatically plot climate models projections and useful statistics.
}
\examples{
fpath <- system.file("extdata/", package="chatR")
exmp <- load_data(country = "Moldova", variable="hurs", n.cores=6,
              path.to.rcps = fpath) \%>\%
projections(., season = 1:12)


}
