% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projections}
\alias{projections}
\title{Analysis of future projections}
\usage{
projections(
  data,
  bias.correction = F,
  uppert = NULL,
  lowert = NULL,
  season,
  consecutive = F,
  n.sessions = 1,
  duration = "max"
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{numeric, seasons to select. For example, 1:12}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{n.sessions}{numeric, number of sessions to use, default is one. Parallelization can be useful when multiple scenarios are used (RCPS, SSPs). However, note that parallelizing will increase RAM usage}

\item{duration}{character, either "max" or "total". Used only when consecutive is TRUE}
}
\value{
list with SpatRaster. .[[1]] contains SpatRaster for the ensemble mean. .[[2]] contains SpatRaster for the ensemble sd and .[[3]] conins SpatRaster for individual models
}
\description{
Automatically process climate model projections and compute useful statistics
}
