% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_two.R
\name{projections}
\alias{projections}
\title{Step two: projection analysis}
\usage{
projections(
  data,
  bias.correction = F,
  uppert = NULL,
  lowert = NULL,
  season,
  scaling.type = "additive",
  consecutive = F,
  duration = "max"
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{Numerical, seasons to select. For example, 1:12}

\item{scaling.type}{character, default to "additive". Indicates whether to use multiplicative or additive approach for bias correction}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{character, either "max" or "total".}
}
\value{
list with raster stacks
}
\description{
Automatically process climate model projections and compute useful statistics.
}
\examples{
load_data(country = "Somalia", variable="tas", years.hist=2000, years.proj=2010,
              path.to.data = "CORDEX-CORE", domain="AFR-22") \%>\%
projections(., season = 1:12)


}
