% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{projections}
\alias{projections}
\title{Climate projections visualization}
\usage{
projections(
  data,
  bias.correction = F,
  uppert = NULL,
  lowert = NULL,
  legend.range = NULL,
  trends = F,
  season,
  palette = NULL,
  method = "scaling",
  prov.country = NULL,
  consecutive = FALSE,
  duration = "max"
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{legend.range}{numeric of length=2,}

\item{trends}{logical, if TRUE applies linear regression}

\item{season}{Numerical, seasons to select. For example, 1:12}

\item{palette}{character vector indicating colors to be used}

\item{method}{character, default to "scaling". Indicates the bias correction method}

\item{prov.country}{character, country for which provinces are highlighted}

\item{consecutive}{logical, to use in conjuunction with lowert or uppert}

\item{duration}{character, either "max" or "total".}
}
\value{
ggplot object
}
\description{
Automatically plot climate models projections and useful statistics.
}
\examples{
fpath <- system.file("extdata/", package="chatR")
exmp <- load_data(country = "Moldova", variable="hurs", n.cores=6,
              path.to.rcps = fpath)
projections(exmp, season = 1:12)


}
