% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{remove_facets}
\alias{remove_facets}
\title{Clean text from plots}
\usage{
remove_facets(p, position = c("both", "x", "y"))
}
\arguments{
\item{p}{Outout of plotting function.}

\item{position}{Character; which facet strips to remove. One of "both", "x", or "y".}
}
\value{
A ggplot object with the specified strip texts removed.
}
\description{
Remove facet labels from a ggplot object.
}
