% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends}
\alias{trends}
\title{Analysis of long-term trends}
\usage{
trends(
  data,
  bias.correction = FALSE,
  uppert = NULL,
  lowert = NULL,
  season,
  consecutive = FALSE,
  duration = "max",
  frequency = F,
  intraannual_var = FALSE,
  observation = FALSE,
  n.sessions = 1
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical, whether to perform bias.correction or not}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{list, containing seasons to select. For example, list(1:6, 7:12)}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{A parameter that can be set to either "max" or a specific number. It is relevant only when 'consecutive' is set to TRUE. For instance, to calculate the count of consecutive days with Tmax (maximum temperature) above 35°C, lasting for more than 3 days, you can set 'uppert' to 35, 'consecutive' to TRUE, and 'duration' to 3.}

\item{frequency}{logical value. This parameter is relevant only when 'consecutive' is set to TRUE and 'duration' is not set to "max". For instance, if you want to determine the count of heatwaves, defined as the number of days with Tmax (maximum temperature) exceeding 35°C for a minimum of 3 consecutive days, set 'uppert' to 35, 'consecutive' to TRUE, 'duration' to 3, and 'frequency' to TRUE.}

\item{intraannual_var}{logical, whether linear regression is applied to annual variability, measured as standard deviation}

\item{observation}{logical, whether to visualize trends for the observational dataset or projections}

\item{n.sessions}{numeric, number of sessions to use, default is one. Parallelisation can be useful when multiple scenarios are used (RCPS, SSPs). However, note that parallelising will increase RAM usage}
}
\value{
list with SpatRaster. To explore the output run attributes(output)
}
\description{
Compute multivariate linear regression and linear regression through design-based inference
}
