% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends}
\alias{trends}
\title{Analysis of long-term trends}
\usage{
trends(
  data,
  bias.correction = FALSE,
  uppert = NULL,
  lowert = NULL,
  scaling.type = "additive",
  season,
  consecutive = FALSE,
  duration = "max",
  interannual_var = FALSE,
  historical = FALSE
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical, whether to perform bias.correction or not}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{scaling.type}{character, default to "additive". Indicates whether to use multiplicative or additive approach for bias correction.}

\item{season}{numeric, seasons to select. For example, 1:12}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{character, either "max" or "total"}

\item{interannual_var}{logical, whether linear regression is applied to annual variability, measured as standard deviation}

\item{historical}{logical, whether to visualize trends for the historical period or projections}
}
\value{
list with raster stacks. To explore the output run attributes(output)
}
\description{
Compute multivariate linear regression and linear regression through design-based inference
}
\examples{
exmp <- load_data(country = "Sudan", variable="pr", years.hist=1995, years.proj=2030:2060,
path.to.data = "CORDEX-CORE", path.to.obs="W5E5", domain="AFR-22", aggr.m="sum") \%>\%
trends(., season = 1:12, historical=F, interannual_var=T)

}
