% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends}
\alias{trends}
\title{Analysis of long-term trends}
\usage{
trends(
  data,
  bias.correction = FALSE,
  uppert = NULL,
  lowert = NULL,
  season,
  consecutive = FALSE,
  duration = "max",
  intraannual_var = FALSE,
  historical = FALSE,
  n.sessions = 1
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical, whether to perform bias.correction or not}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{season}{numeric, seasons to select. For example, 1:12}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{character, either "max" or "total"}

\item{intraannual_var}{logical, whether linear regression is applied to annual variability, measured as standard deviation}

\item{historical}{logical, whether to visualize trends for the historical period or projections}

\item{n.sessions}{numeric, number of sessions to use, default is one. Parallelisation can be useful when multiple scenarios are used (RCPS, SSPs). However, note that parallelising will increase RAM usage}
}
\value{
list with SpatRaster. To explore the output run attributes(output)
}
\description{
Compute multivariate linear regression and linear regression through design-based inference
}
\examples{
exmp <- load_data(country = "Togo", variable="pr", years.hist=1995, years.proj=2030:2055,
path.to.data = "CORDEX-CORE",domain="AFR-22", aggr.m="sum")
trd_exmp <- trends(exmp, season = 1:12, historical=F, intraannual_var=F)
plotting(trd_exmp, ensemble=F, plot_titles = "°C/year", frequencies=F, spatial_aggr=F)

}
