% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends}
\alias{trends}
\title{trends analysis}
\usage{
trends(
  data,
  bias.correction = FALSE,
  uppert = NULL,
  lowert = NULL,
  scaling.type = "additive",
  season,
  consecutive = FALSE,
  duration = "max",
  historical = FALSE
)
}
\arguments{
\item{data}{output of load_data}

\item{bias.correction}{logical, whether to perform bias.correction or not}

\item{uppert}{numeric of length=1, upper threshold}

\item{lowert}{numeric of length=1, lower threshold}

\item{scaling.type}{character, default to "additive". Indicates whether to use multiplicative or additive approach for bias correction.}

\item{season}{Numerical, seasons to select. For example, 1:12}

\item{consecutive}{logical, to use in conjunction with lowert or uppert}

\item{duration}{character, either "max" or "total".}

\item{historical}{logical, whether to visualize trends for the historical period or projections}
}
\value{
list with raster stacks
}
\description{
Perform linear trends.
}
\examples{
fpath <- system.file("extdata/", package="CAVAanalytics")
exmp <- load_data(country = "Moldova", variable="hurs", years.hist=2000, years.proj=2010, path.to.data = fpath) \%>\%
projections(., season = 1:12)


}
